enhancement - Object(need not define the keys)
we need not define a key for an object. if the key and the value are same, we can just ignore the key. so in this case if we just define the value that is enough. Even in function, we dont have to define the function. we just say printDetails(){}

in a nutshell, we need not define the keys for function and those values whose key and values if they are same.