excercise-gallery
used navbar with a small change which says navbar-inverse with which i got the black and white navbar but doesnt really matter coz i changed the colors. And then i also used navbar-fixed-top which is how the navbar is fixed on top. used grid system down the page and used jumbotron and introduced glyphicons which are built in bootstrap icons which are relatively limited. Another class we used is "thumbnail" which we can put around the image to constrain the image into a grid and add border. Lastly used specificity and changed the background color,color of font, links and nav brand.